requirements.txt
sentence-transformers
faiss-cpu
scikit-learn
numpy
jupyterlab
flask